nemo-guardrails/requirements.txt
nemoguardrails>=0.5.0
langchain>=0.0.325
langchain-openai>=0.0.2
python-dotenv>=1.0.0
openai>=1.3.5
chromadb>=0.4.15
